+++ internal/services/filesharer/internal/handler/local_file.go
@@ -11,6 +11,7 @@ import (
 	"github.com/number571/hidden-lake/internal/services/filesharer/internal/utils"
 	hls_settings "github.com/number571/hidden-lake/internal/services/filesharer/pkg/settings"
 	"github.com/number571/hidden-lake/internal/utils/api"
+	"github.com/number571/hidden-lake/internal/utils/chars"
 	http_logger "github.com/number571/hidden-lake/internal/utils/logger/http"
 	hlk_client "github.com/number571/hidden-lake/pkg/api/kernel/client"
 )
@@ -32,11 +33,11 @@ func HandleLocalFileAPI(
 
 		queryParams := pR.URL.Query()
 		aliasName := queryParams.Get("friend")
-		fileName := filepath.Base(queryParams.Get("name"))
 
-		if fileName == "" || fileName != queryParams.Get("name") {
-			pLogger.PushWarn(logBuilder.WithMessage("got_another_name"))
-			_ = api.Response(pW, http.StatusBadRequest, "failed: got another name")
+		fileName := filepath.Base(queryParams.Get("name"))
+		if fileName == "" || chars.HasNotGraphicCharacters(fileName) || fileName != queryParams.Get("name") {
+			pLogger.PushWarn(logBuilder.WithMessage("got_invalid_name"))
+			_ = api.Response(pW, http.StatusBadRequest, "failed: got invalid name")
 			return
 		}